requirements.txt
supabase
streamlit
g4f
pytz
requests
beautifulsoup4
pandas
uuid
googlesearch-python
timezonefinder
langdetect
arxiv
diskcache
curl_cffi>=0.5.0
konlpy
soynlp
pypdf>=3.0.0
